+++ src/main/java/com/miljanilic/catalog/exception/FailedParsingSchemaException.java
@@ -1,4 +1,4 @@
-package com.miljanilic.catalog;
+package com.miljanilic.catalog.exception;
 
 public class FailedParsingSchemaException extends RuntimeException {
